C:\SlowControlsMaggie\EPICS\Utilities\tcIoc: Adding support for Windows x64
git-svn-id: https://redoubt.ligo-wa.caltech.edu/svn/slowcontrols/branches/maggie/EPICS/Utilities/tcIoc@4633 2d2f3c23-3d3d-4554-b2b7-a247361b32a4

+++ ParseTpy.cpp
@@ -361,7 +361,7 @@ void type_map::insert (value_type val)
  ************************************************************************/
 bool compareNamesWoNamespace (const std::stringcase& p1, const std::stringcase& p2)
 {
-	int pos = p1.length() - p2.length();
+	int pos = static_cast<int>(p1.length() - p2.length());
 	if (pos > 0) {
 		return (p1[pos-1] == '.') &&
 			(p1.compare (pos, std::stringcase::npos, p2) == 0);
